address @illicitonion comments
