+++ .github/workflows/buggy-app.yml
@@ -10,7 +10,6 @@ jobs:
       - uses: actions/checkout@v3
       - name: Set up Go
         uses: actions/setup-go@v3
-        # Remove due to "There are no cache folders on the disk" error
         with:
           go-version-file: "buggy-app/go.mod"
           cache-dependency-path: "buggy-app/go.sum"
