+++ buggy-app/auth/client_test.go
@@ -12,6 +12,30 @@ import (
 	"google.golang.org/grpc"
 )
 
+// Internal grpcAuthService struct that implements the gRPC server interface
+// for testing
+type mockGrpcAuthService struct {
+	pb.UnimplementedAuthServer
+
+	result *pb.Result
+	err    error
+
+	Calls int
+}
+
+func newMockGrpcService(result *pb.Result, err error) *mockGrpcAuthService {
+	return &mockGrpcAuthService{
+		result: result,
+		err:    err,
+	}
+}
+
+// Verify checks a Input for authentication validity
+func (as *mockGrpcAuthService) Verify(ctx context.Context, in *pb.Input) (*pb.Result, error) {
+	as.Calls += 1
+	return as.result, as.err
+}
+
 func TestClientCreate(t *testing.T) {
 	config := Config{
 		Port: 8010,
